+++ curtains/middleware.py
@@ -1,0 +1,10 @@
+from __future__ import unicode_literals
+
+from .basic_auth import basic_auth
+from .only_staff import only_staff
+
+
+__all__ = (
+    'basic_auth',
+    'only_staff',
+)