+++ CHANGELOG.rst
@@ -4,12 +4,15 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+- Renamed the project to django-curtains, and added a basic
+  authorization-using middleware.
+
 
 `0.1`_ (2018-02-26)
 ~~~~~~~~~~~~~~~~~~~
 
 - Initial public version.
 
-.. _0.1: https://github.com/matthiask/django-only-staff/commit/89bb93c5cdba
-.. _0.2: https://github.com/matthiask/django-only-staff/compare/0.1...0.2
-.. _Next version: https://github.com/matthiask/django-only-staff/compare/0.1...master
+.. _0.1: https://github.com/matthiask/django-curtains/commit/89bb93c5cdba
+.. _0.2: https://github.com/matthiask/django-curtains/compare/0.1...0.2
+.. _Next version: https://github.com/matthiask/django-curtains/compare/0.1...master